refactor(ensure): port enum tests to ts
